feat: refactor chatBox component into composables
